add u64

+++ ctrtool/main.c
@@ -83,17 +83,17 @@ static u32 align(u32 offset, u32 alignment)
 	return (offset + (alignment-1)) & mask;
 }
 
-unsigned long long getle64(const u8* p)
+u64 getle64(const u8* p)
 {
-	unsigned long long n = p[0];
+	u64 n = p[0];
 
-	n |= (unsigned long long)p[1]<<8;
-	n |= (unsigned long long)p[2]<<16;
-	n |= (unsigned long long)p[3]<<24;
-	n |= (unsigned long long)p[4]<<32;
-	n |= (unsigned long long)p[5]<<40;
-	n |= (unsigned long long)p[6]<<48;
-	n |= (unsigned long long)p[7]<<56;
+	n |= (u64)p[1]<<8;
+	n |= (u64)p[2]<<16;
+	n |= (u64)p[3]<<24;
+	n |= (u64)p[4]<<32;
+	n |= (u64)p[5]<<40;
+	n |= (u64)p[6]<<48;
+	n |= (u64)p[7]<<56;
 	return n;
 }
